requirements.txt
Django==3.2.13
djangorestframework==3.13.1
django_filter==22.1
openpyxl==3.0.10
django_rest_multiple_models==2.1.3